move mstest runtime provider
