+++ TechTalk.SpecFlow/PlatformSpecific/UnitTestProviders.cs
@@ -9,7 +9,7 @@ namespace TechTalk.SpecFlow.Infrastructure
         {
             container.RegisterTypeAs<NUnitRuntimeProvider, IUnitTestRuntimeProvider>("nunit");
             container.RegisterTypeAs<XUnitRuntimeProvider, IUnitTestRuntimeProvider>("xunit");
-            container.RegisterTypeAs<MsTestRuntimeProvider, IUnitTestRuntimeProvider>("mstest");
+            
         }
     }
 }
